add message> Success / add usecase

+++ controllers/home/home.php
@@ -15,7 +15,7 @@ $manager = new AccountManager($db);
 **Add account
 */
 if (isset($_POST['addAccount']) AND isset($_POST['name'])) {
-  if (!empty($_POST['name'])) {
+  if (!empty($_POST['addAccount']) AND !empty($_POST['name'])) {
     // Security XSS
     $name=htmlspecialchars($_POST['name']);
 
@@ -33,26 +33,28 @@ if (isset($_POST['addAccount']) AND isset($_POST['name'])) {
 /*
 **Supp account
 */
-if (isset($_POST['suppAccount'])) {
-  // Security XSS
-  $id=sercureXSS($_POST['id']);
+if (isset($_POST['suppAccount']) && isset($_POST['id'])) {
+  if (!empty($_POST['suppAccount']) && !empty($_POST['id'])) {
+    // Security XSS
+    $id=sercureXSS($_POST['id']);
 
-  // delete on DDB
-  $manager->delete($id);
+    // delete on DDB
+    $manager->delete($id);
+  }
 }
 
 /*
 **Output money
 */
-if (isset($_POST['outputMoney']) AND isset($_POST['amountOutput']) AND isset($_POST['solde'])) {
-  if (!empty($_POST['amountOutput']) AND !empty($_POST['solde'])){
+if (isset($_POST['outputMoney']) AND isset($_POST['amountOutput']) AND isset($_POST['sold'])) {
+  if (!empty($_POST['outputMoney']) AND !empty($_POST['amountOutput']) AND !empty($_POST['sold'])){
     // Security XSS
     $id=sercureXSS($_POST['id']);
     $amount=sercureXSS($_POST['amountOutput']);
-    $solde=sercureXSS($_POST['solde']);
+    $sold=sercureXSS($_POST['sold']);
 
     // creation object outputMoneyAccount
-    $outputMoneyAccount=new account(['solde'=>$solde,'id'=>$id]);
+    $outputMoneyAccount=new account(['sold'=>$sold,'id'=>$id]);
 
     // use adding money methode
     $message=$outputMoneyAccount->outputMoney($amount);
@@ -60,7 +62,6 @@ if (isset($_POST['outputMoney']) AND isset($_POST['amountOutput']) AND isset($_P
     // update on DDB
     $manager->update($outputMoneyAccount);
 
-    $message="Success";
   }
     // if input empty
   else {
@@ -71,15 +72,15 @@ if (isset($_POST['outputMoney']) AND isset($_POST['amountOutput']) AND isset($_P
 /*
 **Add money
 */
-if (isset($_POST['addMoney']) AND isset($_POST['amountAdd']) AND isset($_POST['solde']) AND isset($_POST['id'])) {
-  if (!empty($_POST['amountAdd'])){
+if (isset($_POST['addMoney']) AND isset($_POST['amountAdd']) AND isset($_POST['sold']) AND isset($_POST['id'])) {
+  if (!empty($_POST['addMoney']) AND !empty($_POST['amountAdd']) AND !empty($_POST['sold']) AND !empty($_POST['id'])){
     // Security XSS
     $id=sercureXSS($_POST['id']);
     $amount=sercureXSS($_POST['amountAdd']);
-    $solde=sercureXSS($_POST['solde']);
+    $sold=sercureXSS($_POST['sold']);
 
     // creation object addMoneyAccount
-    $addMoneyAccount=new account(['solde'=>$solde,'id'=>$id]);
+    $addMoneyAccount=new account(['sold'=>$sold,'id'=>$id]);
 
     // use adding money methode
     $message=$addMoneyAccount->addMoney($amount);
@@ -87,7 +88,6 @@ if (isset($_POST['addMoney']) AND isset($_POST['amountAdd']) AND isset($_POST['s
     // update on DDB
     $manager->update($addMoneyAccount);
 
-    $message="Success";
   }
     // if input empty
   else {
@@ -100,36 +100,36 @@ if (isset($_POST['addMoney']) AND isset($_POST['amountAdd']) AND isset($_POST['s
 /*
 **Transfert money
 */
-if (isset($_POST['transfertMoney']) && isset($_POST['amountTransfert']) && isset($_POST['id']) && isset($_POST['idAddMoneyAccount']) && isset($_POST['solde'])){
-  if (!empty($_POST['transfertMoney']) && !empty($_POST['amountTransfert']) && !empty($_POST['id']) && !empty($_POST['idAddMoneyAccount']) && !empty($_POST['solde'])){
+if (isset($_POST['transfertMoney']) && isset($_POST['amountTransfert']) && isset($_POST['id']) && isset($_POST['idAddMoneyAccount']) && isset($_POST['sold'])){
+  if (!empty($_POST['transfertMoney']) && !empty($_POST['amountTransfert']) && !empty($_POST['id']) && !empty($_POST['idAddMoneyAccount']) && !empty($_POST['sold'])){
     // Security XSS
     $idAddMoneyAccount=sercureXSS($_POST['idAddMoneyAccount']);
     $idOutputMoneyAccount=sercureXSS($_POST['id']);
-    $soldeOutputMoneyAccount=sercureXSS($_POST['solde']);
+    $soldOutputMoneyAccount=sercureXSS($_POST['sold']);
     $amount=sercureXSS($_POST['amountTransfert']);
 
     // Ouput money on account
       // creation of outputMoneyAccount
-    $outputMoneyAccount=new account(['solde'=>$soldeOutputMoneyAccount,'id'=>$idOutputMoneyAccount]);
+    $outputMoneyAccount=new account(['sold'=>$soldOutputMoneyAccount,'id'=>$idOutputMoneyAccount]);
 
       // use ouput methode
     $validator=$outputMoneyAccount->outputMoney($amount);
 
     // if ouput money its ok
-    if ($validator===true) {
+    if ($validator=="Success") {
       // get addMoneyAccount on DDB (object)
       $addMoneyAccount = $manager->get($idAddMoneyAccount);
 
-      // add solde
+      // add sold
       $validator=$addMoneyAccount->addMoney($amount);
 
       // if addMoney money its ok
-      if ($validator===true) {
+      if ($validator=="Success") {
         // update on DDB
         $manager->update($outputMoneyAccount);
         $manager->update($addMoneyAccount);
 
-        $message="success";
+        $message="Success";
       }
       else {
         $message="can't add money on account";
